# Improve performance of `tuist generate` when cache is used by refactoring CacheGraphMutator