--- a/Sources/TuistCore/GraphTraverser/GraphTraversing.swift
+++ b/Sources/TuistCore/GraphTraverser/GraphTraversing.swift
@@ -39,6 +39,9 @@ public protocol GraphTraversing {
     /// Returns all the targets of the project.
     func allTargets() -> Set<GraphTarget>
 
+    /// Returns all the targets of the project, topological sorted.
+    func allTargetsTopologicalSorted() throws -> [GraphTarget]
+
     /// Returns all the internal targets, that is, excluding `Dependencies`.
     func allInternalTargets() -> Set<GraphTarget>