### Short description 📝

As initially saw in #4145:

> I noticed we would spent lot of time because we don't keep track of dependency that are already visited in CacheGraphMutator. With a large workspace with shared dependencies I noticed a 10x improvement, from 6s to 0.5s.

This PR refactor the logic and change the algorithm we are using for mapping replaceable targets, in short, by visiting the graph using topological order we don't need to keep track of all transitive dependencies but only the direct dependency, simplify the need to keep track of already visited nodes and getting even better performance improvements (6s vs 0.2s).
I also added some documentation and try to have better separation for the methods inside `CacheGraphMutator`.

### How to test the changes locally 🧐

Unit tests and acceptances covers already `CacheGraphMutator`.

### Checklist ✅

- [x] The code architecture and patterns are consistent with the rest of the codebase.
- [x] The changes have been tested following the [documented guidelines](https://docs.tuist.io/contributors/testing-strategy/).
- [x] The `CHANGELOG.md` has been updated to reflect the changes. In case of a breaking change, it's been flagged as such.
- [x] In case the PR introduces changes that affect users, the documentation has been updated.
- [x] In case the PR introduces changes that affect how the cache artifact is generated starting from the `TuistGraph.Target`, the `Constants.cacheVersion` has been updated.